update: libft

+++ project/libft/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: shachowd <shachowd@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/22 13:02:57 by shachowd          #+#    #+#             */
-/*   Updated: 2024/05/16 12:15:24 by shachowd         ###   ########.fr       */
+/*   Updated: 2024/05/16 15:18:05 by shachowd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int					ft_isdigit(int c);
 int					ft_isalnum(int c);
 int					ft_isascii(int c);
 int					ft_isprint(int c);
-int					ft_strlen(const char *s);
+size_t				ft_strlen(const char *s);
 void				*ft_memset(void *str, int c, size_t len);
 void				ft_bzero(void *s, size_t n);
 void				*ft_memcpy(void *dst, const void *src, size_t n);
@@ -46,6 +46,7 @@ char				*ft_strnstr(const char *haystack,
 int					ft_atoi(char const *str);
 void				*ft_calloc(size_t count, size_t size);
 char				*ft_strdup(const char *src);
+
 
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_substr(char const *s, unsigned int start,
